quart auth introduction checkpoint (#66)
* quart auth introduction checkpoint

* quart auth introduction checkpoint 2/?

* always at least one active admin

* fine grain user permissions

* refactor; remove delta for sql placeholders

* ui titles

---------

+++ src/blueprints/bp_moderation.py
@@ -3,11 +3,12 @@ from html import escape
 
 from flask_paginate import Pagination
 from quart import Blueprint, abort, flash, jsonify, redirect, request, url_for
+from quart_auth import current_user, login_required
 
 from asagi_converter import get_post, move_post_to_delete_table
 from boards import board_shortnames
 from configs import mod_conf
-from enums import AuthActions, ModStatus, PublicAccess
+from enums import ModStatus, PublicAccess
 from forms import ReportUserForm
 from leafs import (
     generate_post_html,
@@ -15,7 +16,6 @@ from leafs import (
     post_files_hide,
     post_files_show
 )
-from moderation.auth import auth, authorization_required
 from moderation.filter_cache import fc
 from moderation.report import (
     create_report,
@@ -25,6 +25,12 @@ from moderation.report import (
     get_report_count_all,
     get_report_count_f,
     get_reports_f
+)
+from moderation.user import (
+    Permissions,
+    require_is_active,
+    require_is_admin,
+    require_permissions
 )
 from render import render_controller
 from templates import template_reports_index
@@ -142,7 +148,9 @@ async def make_report_pagination(mod_status: ModStatus, board_shortnames: list[s
 
 @bp.get('/reports/closed')
 @bp.get('/reports/closed/<int:page_num>')
-@authorization_required
+@login_required
+@require_is_active
+@require_permissions([Permissions.report_read])
 async def reports_closed(page_num: int=0):
     page_size = 20
     reports = await get_reports_f(mod_status=ModStatus.closed, board_shortnames=board_shortnames, page_num=page_num, page_size=page_size)
@@ -154,14 +162,16 @@ async def reports_closed(page_num: int=0):
         reports=await formulate_reports_for_html_table(reports),
         title='Closed Reports',
         tab_title='Closed Reports',
-        is_logged_in=True,
-        is_admin=await auth(AuthActions.is_admin),
+        is_authenticated=True,
+        is_admin=current_user.is_admin,
     )
 
 
 @bp.get('/reports/open')
 @bp.get('/reports/open/<int:page_num>')
-@authorization_required
+@login_required
+@require_is_active
+@require_permissions([Permissions.report_read])
 async def reports_open(page_num: int=0):
     page_size=20
     reports = await get_reports_f(mod_status=ModStatus.open, board_shortnames=board_shortnames, page_num=page_num, page_size=page_size)
@@ -174,8 +184,8 @@ async def reports_open(page_num: int=0):
         reports=await formulate_reports_for_html_table(reports),
         title='Reports',
         tab_title='Reports',
-        is_logged_in=True,
-        is_admin=await auth(AuthActions.is_admin),
+        is_authenticated=True,
+        is_admin=current_user.is_admin,
     )
 
 
@@ -192,6 +202,9 @@ async def reports_action_routine(report_parent_id: int, action: str, mod_notes: 
 
     match action:
         case 'report_delete':
+            if not current_user.has_permissions([Permissions.report_delete]):
+                return f'Need permissions for {Permissions.report_delete.name}'
+
             report = await delete_report_if_exists(report_parent_id)
             flash_msg = f'Report was already deleted.'
             if report:
@@ -200,40 +213,47 @@ async def reports_action_routine(report_parent_id: int, action: str, mod_notes: 
             return flash_msg
 
         case 'post_delete':
+            if not current_user.has_permissions([Permissions.post_delete]):
+                return f'Need permissions for {Permissions.post_delete.name}'
+            
             # Note: do not delete the report here. It is still needed to filter outgoing posts from full text search.
-            post, result = await move_post_to_delete_table(report.board_shortname, report.num)
-            if result == 0:
-                flash_msg = 'Did not locate post in asagi database. Did nothing as a result.'
-            elif result == -1:
-                flash_msg = 'Did not transfer post to asagi\'s delete table. It is still in the board table.'
-            elif result == 1:
-                flash_msg = 'Post transfered to asagi\'s delete table. It is no longer in the board table.'
-            else:
-                raise ValueError(post, result)
+            post, flash_msg = await move_post_to_delete_table(report.board_shortname, report.num)
 
             full_del, prev_del = post_files_delete(post)
             flash_msg += ' Deleted full media.' if full_del else ' Did not delete full media.'
             flash_msg += ' Deleted thumbnail.' if prev_del else ' Did not delete thumbnail.'
 
         case 'media_delete':
+            if not current_user.has_permissions([Permissions.media_delete]):
+                return f'Need permissions for {Permissions.media_delete.name}'
+            
             post = await get_post(report.board_shortname, report.num)
             full_del, prev_del = post_files_delete(post)
             flash_msg += ' Deleted full media.' if full_del else ' Did not delete full media.'
             flash_msg += ' Deleted thumbnail.' if prev_del else ' Did not delete thumbnail.'
 
         case 'media_hide':
+            if not current_user.has_permissions([Permissions.media_hide]):
+                return f'Need permissions for {Permissions.media_hide.name}'
+
             post = await get_post(report.board_shortname, report.num)
             full_hid, prev_hid = post_files_hide(post)
             flash_msg += ' Hid full media.' if full_hid else ' Did not hide full media.'
             flash_msg += ' Hid thumbnail.' if prev_hid else ' Did not hide thumbnail.'
 
         case 'media_show':
+            if not current_user.has_permissions([Permissions.media_show]):
+                return f'Need permissions for {Permissions.media_show.name}'
+            
             post = await get_post(report.board_shortname, report.num)
             full_sho, prev_sho = post_files_show(post)
             flash_msg += ' Showing full media.' if full_sho else ' Did not reveal full media.'
             flash_msg += ' Showing thumbnail.' if prev_sho else ' Did not reveal thumbnail.'
 
         case 'post_show':
+            if not current_user.has_permissions([Permissions.post_show]):
+                return f'Need permissions for {Permissions.post_show.name}'
+            
             report = await edit_report_if_exists(report_parent_id, public_access=PublicAccess.visible)
             if report:
                 await fc.delete_post(report['board_shortname'], report['num'], report['op'])
@@ -246,6 +266,9 @@ async def reports_action_routine(report_parent_id: int, action: str, mod_notes: 
                 flash_msg += ' Showing thumbnail.' if prev_sho else ' Did not reveal thumbnail.'
 
         case 'post_hide':
+            if not current_user.has_permissions([Permissions.post_hide]):
+                return f'Need permissions for {Permissions.post_hide.name}'
+
             report = await edit_report_if_exists(report_parent_id, public_access=PublicAccess.hidden)
             if report:
                 await fc.insert_post(report['board_shortname'], report['num'], report['op'])
@@ -258,14 +281,23 @@ async def reports_action_routine(report_parent_id: int, action: str, mod_notes: 
                 flash_msg += ' Hid thumbnail.' if prev_hid else ' Did not hide thumbnail.'
 
         case 'report_close':
+            if not current_user.has_permissions([Permissions.report_close]):
+                return f'Need permissions for {Permissions.report_close.name}'
+
             report = await edit_report_if_exists(report_parent_id, mod_status=ModStatus.closed)
             flash_msg = 'Report moved to closed reports.'
 
         case 'report_open':
+            if not current_user.has_permissions([Permissions.report_open]):
+                return f'Need permissions for {Permissions.report_open.name}'
+
             report = await edit_report_if_exists(report_parent_id, mod_status=ModStatus.open)
             flash_msg = 'Report moved to opened reports.'
 
         case 'report_save_notes':
+            if not current_user.has_permissions([Permissions.report_save_notes]):
+                return f'Need permissions for {Permissions.report_save_notes.name}'
+
             # falsey mod_notes are valid
             await edit_report_if_exists(report_parent_id, mod_notes=mod_notes)
             flash_msg = f'Report had their moderation notes saved.'
@@ -277,7 +309,9 @@ async def reports_action_routine(report_parent_id: int, action: str, mod_notes: 
 
 
 @bp.route('/reports/<int:report_parent_id>/<string:action>', methods=['POST'])
-@authorization_required
+@login_required
+@require_is_active
+@require_permissions([Permissions.report_read])
 async def reports_action(report_parent_id: int, action: str):
     form = (await request.form)
     redirect_endpoint = form.get('endpoint', 'bp_moderation.reports_open')
@@ -290,7 +324,9 @@ async def reports_action(report_parent_id: int, action: str):
 
 
 @bp.route('/reports/bulk/<string:action>', methods=['POST'])
-@authorization_required
+@login_required
+@require_is_active
+@require_permissions([Permissions.report_read])
 async def reports_action_bulk(action: str):
     data = (await request.get_json())
